let_service/requirements-let-service.txt
flask>=3.0.0
flask_sqlalchemy>=3.1.1
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pymysql>=1.1.0
cryptography>=41.0.0
